+++ core/pom.xml
@@ -43,12 +43,19 @@
 			<version>4.6.4</version>
 		</dependency>
 
-		<!-- https://mvnrepository.com/artifact/com.mesosphere/marathon-client -->
 		<dependency>
 			<groupId>com.mesosphere</groupId>
 			<artifactId>marathon-client</artifactId>
 			<version>0.6.2</version>
 		</dependency>
+
+		<dependency>
+			<groupId>org.assertj</groupId>
+			<artifactId>assertj-core</artifactId>
+			<version>3.14.0</version>
+			<scope>test</scope>
+		</dependency>
+
 
 	</dependencies>
