+++ core/src/test/java/fr/insee/innovation/marathontokubernetes/core/services/marathon/MarathonImporterTest.java
@@ -1,6 +1,7 @@
 package fr.insee.innovation.marathontokubernetes.core.services.marathon;
 
 import java.io.InputStream;
+import java.util.List;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -9,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
 import mesosphere.marathon.client.model.v2.App;
+import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
 public class MarathonImporterTest {
@@ -20,8 +22,8 @@ public class MarathonImporterTest {
     @ValueSource(strings = {"/marathon/drawio.json"})
     public void shouldImportDrawIo(String location) {
         InputStream input = getClass().getResourceAsStream(location);
-        App app = importer.importMarathonApp(input);
-        Assertions.assertNotNull(app);
-        Assertions.assertEquals(2, app.getCpus());
+        List<App> apps = importer.importMarathonApp(input);
+        assertThat(apps).isNotNull().hasSize(1);
+        assertThat(apps).first().satisfies(app -> assertThat(app.getCpus() == 2));
     }
 }
